// workspaces/ballerina/ballerina-side-panel/src/components/Form/types.ts
export type FormField = {
    key: string;
    label: string;
    type: null | string;
    optional: boolean;
    advanced?: boolean;
    editable: boolean;
    placeholder?: string;
    documentation: string;
    value: string | any[];
    valueType?: string;
    diagnostics?: DiagnosticMessage[];
    items?: string[];
    choices?: PropertyModel[];
    paramManagerProps?: ParamConfig;
    valueTypeConstraint: string | string[];
    groupNo?: number;
    groupName?: string;
    addNewButton?: boolean;
    enabled?: boolean;
    lineRange?: LineRange;
};